First step of abstraction for runner test
