+++ core/src/commonTest/kotlin/com/github/jcornaz/kwik/runner/CustomException.kt
@@ -1,3 +1,5 @@
+package com.github.jcornaz.kwik.runner
+
 /**
  * Dummy exception to be thrown in test and make sure that the SUT don't wrap the exceptions
  */
